pyproject.toml
[project]
name = "gmail-mcp-server"
version = "0.1.0"
description = "Server for Gmail MCP"
readme = "README.md"
authors = [
    { name = "Ivan Lopez Hdez", email = "info@ivanlopezdev.es" }
]
license = { text = "MIT" }
requires-python = ">=3.11"
dependencies = [
    "dotenv>=0.9.9",
    "google-api-python-client>=2.167.0",
    "google-auth-httplib2>=0.2.0",
    "google-auth-oauthlib>=1.2.1",
    "mcp[cli]>=1.6.0",
]

[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project.scripts]
gmail-mcp-server = "gmail_mcp_server.main:main"

[tool.setuptools]
package-dir = {"" = "src"}
packages = ["gmail_mcp_server", "gmail"]
